# openapi.yaml
openapi: '3.1.0'
info:
  title: 'Mission Control API'
  version: '1.0.0'
  description: |-
    Generated from the Backend’s Fastify route table by `pnpm api:spec` (F5.1). Do not edit
    by hand: `pnpm api:spec:check` fails when this file and the routes disagree, and
    `pnpm api:types:check` fails when apps/frontend/src/lib/api/types.generated.ts does.

    REQUEST shapes are the schemas Ajv actually enforces, so they are exact.

    RESPONSE shapes are the schemas each route declares under `schema.response`. Read the
    guarantee precisely: they are DECLARATIONS, NOT SERIALIZERS. The Backend installs a
    serializer compiler that ignores them and calls JSON.stringify, so a schema here can never
    remove a field the handler produced — the failure mode a Fastify response schema normally
    carries. What keeps them true instead is a conformance check that validates every reply
    against its own schema in the test tiers, in both directions: an undeclared property and a
    missing required property are both failures. See apps/backend/src/http/response-schema.ts.

    Operations still marked `x-mc-response-schema: undeclared` are the ones with no declaration,
    and the marker exists so that gap stays countable. Today they are: GET /health,
    GET /services/health and GET /schedule (owned by modules outside this change),
    GET|PUT /settings/{category} (the GET serves one of five different documents, which needs a
...
    and GET /ws.

...
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  '/agent-teams/{id}':
    get:
      operationId: getAgentTeamsById
      summary: 'GET /agent-teams/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentTeam'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    patch:
      operationId: patchAgentTeamsById
      summary: 'PATCH /agent-teams/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type:
                    - string
                    - 'null'
                  maxLength: 2000
                agentIds:
                  type: array
                  maxItems: 50
                  items:
                    type: string
                    pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                projectIds:
                  type: array
                  maxItems: 100
                  items:
                    type: string
                    pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentTeam'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    delete:
      operationId: deleteAgentTeamsById
      summary: 'DELETE /agent-teams/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '204':
          description: 'No Content'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  /agent-workflow-runs:
    get:
      operationId: getAgentWorkflowRuns
      summary: 'GET /agent-workflow-runs'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum:
              - running
              - completed
              - halted
              - stopped
        - name: workflowId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentWorkflowRun'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    post:
      operationId: postAgentWorkflowRuns
      summary: 'POST /agent-workflow-runs'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workflowId
                - projectId
                - task
                - workingDirectory
              additionalProperties: false
              properties:
                workflowId:
                  type: string
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                projectId:
                  type: string
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                task:
                  type: string
                  minLength: 1
                  maxLength: 20000
                workingDirectory:
                  type: string
                  minLength: 1
                  maxLength: 4096
                repositoryId:
                  type: string
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                branch:
                  type: string
                  minLength: 1
                  maxLength: 255
                model:
                  type: string
                  minLength: 1
                  maxLength: 100
                maxSessions:
                  type: integer
                  minimum: 1
                  maximum: 20
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentWorkflowRun'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/agent-workflow-runs/{id}':
    get:
      operationId: getAgentWorkflowRunsById
      summary: 'GET /agent-workflow-runs/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentWorkflowRun'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/agent-workflow-runs/{id}/resume':
    post:
      operationId: postAgentWorkflowRunsByIdResume
      summary: 'POST /agent-workflow-runs/{id}/resume'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type:
                - object
                - 'null'
              additionalProperties: false
              properties: {}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentWorkflowRun'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/agent-workflow-runs/{id}/stop':
    post:
      operationId: postAgentWorkflowRunsByIdStop
      summary: 'POST /agent-workflow-runs/{id}/stop'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type:
                - object
                - 'null'
              additionalProperties: false
              properties: {}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/AgentWorkflowRun'
                  meta:
                    type: object
                    additionalProperties: false
                    required:
                      - stoppedSession
                    properties:
                      stoppedSession:
                        anyOf:
                          - $ref: '#/components/schemas/StoppedSession'
                          - type: 'null'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  /agent-workflows:
    get:
      operationId: getAgentWorkflows
      summary: 'GET /agent-workflows'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum:
              - global
              - project
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentWorkflow'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    post:
      operationId: postAgentWorkflows
      summary: 'POST /agent-workflows'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - scope
                - steps
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type:
                    - string
                    - 'null'
                  maxLength: 2000
                scope:
                  type: string
                  enum:
                    - global
                    - project
                projectId:
                  type:
                    - string
                    - 'null'
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                steps:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items:
                    type: object
                    required:
                      - agentId
                    additionalProperties: false
                    properties:
                      agentId:
                        type: string
                        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                      instructions:
                        type:
                          - string
                          - 'null'
                        maxLength: 4000
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentWorkflow'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
  '/agent-workflows/{id}/cost-estimate':
    get:
      operationId: getAgentWorkflowsByIdCostEstimate
      summary: 'GET /agent-workflows/{id}/cost-estimate'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/WorkflowCostEstimate'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
components:
  schemas:
...
    AgentTeam:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
        - scope
        - projectId
        - members
        - projectIds
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
            - string
            - 'null'
        scope:
          $ref: '#/components/schemas/AgentTeamScope'
        projectId:
          type:
            - string
            - 'null'
        members:
          type: array
          items:
            $ref: '#/components/schemas/AgentTeamMember'
        projectIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ListEnvelopeMeta:
      type: object
      additionalProperties: false
      required:
        - nextCursor
        - limit
      properties:
        nextCursor:
          type:
            - string
            - 'null'
        limit:
          type: integer
...
    StoppedSession:
      type: object
      additionalProperties: false
      required:
        - sessionId
        - outcome
      properties:
        sessionId:
          type: string
        outcome:
          type: string
          enum:
            - ended
            - cancelled
            - already_terminal
...
  responses:
    ErrorEnvelope:
      description: 'Any non-2xx response (F5.4). The status code and the `error.code` registry entry carry the meaning; see docs/tds/04-api-contracts-and-events.md §1.3.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mc_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'mct_ API token'